# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AgentZT Gateway Management API
  description: |
    AgentZT Gateway 管理API提供完整的应用管理、配置管理、统计分析、配额管理和告警管理功能。
    
    ## 认证
    
    所有管理API需要Bearer Token认证。在请求头中添加：
    
    ```
    Authorization: Bearer <your-session-token>
    ```
    
    或使用API Key：
    
    ```
    X-AgentZT-API-Key: <your-api-key>
    ```
    
    ## 应用选择
    
    多应用场景下，可通过以下方式指定应用：
    
    - 请求头：`X-AgentZT-App-ID: <appId>`
    - 请求头：`X-AgentZT-API-Key: <apiKey>`
    - 查询参数：`?appId=<appId>`
    
    ## 错误处理
    
    所有错误响应遵循统一格式：
    
    ```json
    {
      "error": {
        "code": "error_code",
        "message": "Error description",
        "details": {}
      }
    }
    ```
    
    ## 版本
    
    当前API版本为v1，默认路径 `/api` 等同于 `/api/v1`。
  version: 1.0.0
  contact:
    name: AgentZT Team
    email: support@agentzt.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:3000/api
    description: 本地开发服务器
  - url: https://gateway.agentzt.io/api
    description: 生产服务器
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Session Token认证
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-AgentZT-API-Key
      description: API Key认证
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - authentication_error
                - permission_error
                - not_found
                - invalid_request
                - validation_error
                - conflict_error
                - rate_limit_exceeded
                - internal_error
                - service_unavailable
              example: invalid_request
            message:
              type: string
              example: App name must be between 1 and 100 characters
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
              example: req-1699887600000-abc123
            timestamp:
              type: string
              format: date-time
              example: '2024-01-15T10:30:00Z'
...
    AppId:
      type: string
      pattern: '^app-[0-9]+-[a-z0-9]+$'
      example: app-1699887600000-abc123
...
    TimeRange:
      type: string
      enum:
        - day
        - week
        - month
        - year
      description: 时间范围
    Granularity:
      type: string
      enum:
        - hour
        - day
      description: 数据粒度
    StatsOverview:
      type: object
      properties:
        totalDetections:
          type: integer
          description: 总检测次数
          example: 10000
        todayDetections:
          type: integer
          description: 今日检测次数
          example: 150
        passRate:
          type: number
          format: float
          description: 通过率（0-1）
          example: 0.85
        blockRate:
          type: number
          format: float
          description: 阻断率（0-1）
          example: 0.10
        flagRate:
          type: number
          format: float
          description: 标记率（0-1）
          example: 0.05
        avgLatencyMs:
          type: integer
          description: 平均延迟（毫秒）
          example: 150
        totalTokens:
          type: integer
          description: 总Token数量
          example: 500000
    RiskDistribution:
      type: object
      properties:
        byLevel:
          type: object
          properties:
            no_risk:
              type: integer
            low_risk:
              type: integer
            medium_risk:
              type: integer
            high_risk:
              type: integer
          example:
            no_risk: 8500
            low_risk: 500
            medium_risk: 800
            high_risk: 200
        byCategory:
          type: object
          additionalProperties:
            type: integer
          example:
            S1: 150
            S2: 100
            S3: 200
        byApp:
          type: object
          additionalProperties:
            type: integer
          example:
            app-xxx: 5000
            app-yyy: 5000
    TrendDataPoint:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        detections:
          type: integer
        passRate:
          type: number
          format: float
        blockRate:
          type: number
          format: float
    TrendResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TrendDataPoint'
        timeRange:
          $ref: '#/components/schemas/TimeRange'
        granularity:
          $ref: '#/components/schemas/Granularity'

    # ========== Quota Schemas ==========
...
  parameters:
...
    AppIdQueryParam:
      name: appId
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/AppId'
      description: 应用ID（可选）
    TimeRangeParam:
      name: timeRange
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/TimeRange'
      description: 时间范围
    GranularityParam:
      name: granularity
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/Granularity'
      description: 数据粒度
...
paths:
...
  /stats/overview:
    get:
      tags:
        - Stats
      summary: 获取统计概览
      description: 返回指定时间范围内的统计概览数据
      parameters:
        - $ref: '#/components/parameters/TimeRangeParam'
        - $ref: '#/components/parameters/AppIdQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsOverview'
        '400':
          description: 请求参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stats/risk-distribution:
    get:
      tags:
        - Stats
      summary: 获取风险分布
      description: 返回指定时间范围内的风险分布数据
      parameters:
        - $ref: '#/components/parameters/TimeRangeParam'
        - $ref: '#/components/parameters/AppIdQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskDistribution'
        '400':
          description: 请求参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stats/trend:
    get:
      tags:
        - Stats
      summary: 获取趋势数据
      description: 返回指定时间范围和粒度的趋势数据
      parameters:
        - $ref: '#/components/parameters/TimeRangeParam'
        - $ref: '#/components/parameters/GranularityParam'
        - $ref: '#/components/parameters/AppIdQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrendResponse'
        '400':
          description: 请求参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stats/export:
    get:
      tags:
        - Stats
      summary: 导出统计数据
      description: 导出统计数据为CSV或JSON格式
      parameters:
        - name: format
          in: query
          required: false
          schema:
            type: string
            enum:
              - csv
              - json
            default: json
          description: 导出格式
        - $ref: '#/components/parameters/TimeRangeParam'
        - $ref: '#/components/parameters/AppIdQueryParam'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  overview:
                    $ref: '#/components/schemas/StatsOverview'
                  riskDistribution:
                    $ref: '#/components/schemas/RiskDistribution'
                  trend:
                    type: array
                    items:
                      $ref: '#/components/schemas/TrendDataPoint'
                  exportInfo:
                    type: object
                    properties:
                      timeRange:
                        $ref: '#/components/schemas/TimeRange'
                      appId:
                        type: string
                      exportedAt:
                        type: string
                        format: date-time
                      exportedBy:
                        type: string
            text/csv:
              schema:
                type: string
                description: CSV格式统计数据
        '400':
          description: 请求参数无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 无权限
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ========== Quota API ==========
...